Add ability to auto-merge and confirm ballots
This commit implements a combination of the "individual ballots" and
"bypass ballot confirmation" settings, so that on the final submitting
adjudicator to submit a ballot, the "merge ballot" sequence is run and
the resulting ballot is confirmed.

If there is a discrepancy for motions, the ballot becomes "postponed" to
flag the incongruence.

Django 4.2 is required as the DISTINCT clause has been replaced with a
filter on a window.