Update to Django 5
This commit updates the Python version to 3.11 and bumps packages to
their update, most notably Django to version 5. There were some problems
with updating to 3.12 so we'll keep it lower for now.

This new version of Django brings some deprecations, such as model table
indexing being more explicit. Migrations are also needed for this.